add code comments for new functions, move to separate collections file
