+++ internal/ap/activitystreams.go
@@ -17,16 +17,6 @@
 
 package ap
 
-import (
-	"fmt"
-	"net/url"
-	"strconv"
-
-	"github.com/superseriousbusiness/activity/streams"
-	"github.com/superseriousbusiness/activity/streams/vocab"
-	"github.com/superseriousbusiness/gotosocial/internal/paging"
-)
-
 // https://www.w3.org/TR/activitystreams-vocabulary
 const (
 	ActivityAccept          = "Accept"          // ActivityStreamsAccept https://www.w3.org/TR/activitystreams-vocabulary/#dfn-accept
@@ -88,306 +78,3 @@ const (
 	// and https://www.w3.org/TR/activitystreams-vocabulary/#dfn-tag
 	TagHashtag = "Hashtag"
 )
-
-func ToCollectionPageIterator(t vocab.Type) (CollectionPageIterator, error) {
-	switch t := t.(type) {
-	case vocab.ActivityStreamsCollectionPage:
-		return WrapCollectionPage(t), nil
-	case vocab.ActivityStreamsOrderedCollectionPage:
-		return WrapOrderedCollectionPage(t), nil
-	default:
-		return nil, fmt.Errorf("%T(%s) was not CollectionPage-like", t, t.GetTypeName())
-	}
-}
-
-func WrapCollectionPage(page vocab.ActivityStreamsCollectionPage) CollectionPageIterator {
-	return &regularCollectionPageIterator{page, nil}
-}
-
-func WrapOrderedCollectionPage(page vocab.ActivityStreamsOrderedCollectionPage) CollectionPageIterator {
-	return &orderedCollectionPageIterator{page, nil}
-}
-
-type regularCollectionPageIterator struct {
-	vocab.ActivityStreamsCollectionPage
-	items vocab.ActivityStreamsItemsPropertyIterator
-}
-
-func (iter *regularCollectionPageIterator) NextPage() WithIRI {
-	if iter.ActivityStreamsCollectionPage == nil {
-		return nil
-	}
-	return iter.GetActivityStreamsNext()
-}
-
-func (iter *regularCollectionPageIterator) PrevPage() WithIRI {
-	if iter.ActivityStreamsCollectionPage == nil {
-		return nil
-	}
-	return iter.GetActivityStreamsPrev()
-}
-
-func (iter *regularCollectionPageIterator) NextItem() IteratorItemable {
-	if !iter.initItems() {
-		return nil
-	}
-	cur := iter.items
-	iter.items = iter.items.Next()
-	return cur
-}
-
-func (iter *regularCollectionPageIterator) PrevItem() IteratorItemable {
-	if !iter.initItems() {
-		return nil
-	}
-	cur := iter.items
-	iter.items = iter.items.Prev()
-	return cur
-}
-
-func (iter *regularCollectionPageIterator) initItems() bool {
-	if iter.items == nil && iter.ActivityStreamsCollectionPage != nil {
-		items := iter.GetActivityStreamsItems()
-		iter.items = items.Begin()
-	}
-	return (iter.items != nil)
-}
-
-type orderedCollectionPageIterator struct {
-	vocab.ActivityStreamsOrderedCollectionPage
-	items vocab.ActivityStreamsOrderedItemsPropertyIterator
-}
-
-func (iter *orderedCollectionPageIterator) NextPage() WithIRI {
-	if iter.ActivityStreamsOrderedCollectionPage == nil {
-		return nil
-	}
-	return iter.GetActivityStreamsNext()
-}
-
-func (iter *orderedCollectionPageIterator) PrevPage() WithIRI {
-	if iter.ActivityStreamsOrderedCollectionPage == nil {
-		return nil
-	}
-	return iter.GetActivityStreamsPrev()
-}
-
-func (iter *orderedCollectionPageIterator) NextItem() IteratorItemable {
-	if !iter.initItems() {
-		return nil
-	}
-	cur := iter.items
-	iter.items = iter.items.Next()
-	return cur
-}
-
-func (iter *orderedCollectionPageIterator) PrevItem() IteratorItemable {
-	if !iter.initItems() {
-		return nil
-	}
-	cur := iter.items
-	iter.items = iter.items.Prev()
-	return cur
-}
-
-func (iter *orderedCollectionPageIterator) initItems() bool {
-	if iter.items == nil && iter.ActivityStreamsOrderedCollectionPage != nil {
-		items := iter.GetActivityStreamsOrderedItems()
-		iter.items = items.Begin()
-	}
-	return (iter.items != nil)
-}
-
-type CollectionParams struct {
-	// Containing collection
-	// ID (i.e. NOT the page).
-	ID *url.URL
-
-	// Total no. items.
-	Total int
-}
-
-type CollectionPageParams struct {
-	// containing collection.
-	CollectionParams
-
-	// Paging details.
-	Current *paging.Page
-	Next    *paging.Page
-	Prev    *paging.Page
-	Query   url.Values
-
-	// Item appender for each item at index.
-	Append func(int, ItemsPropertyBuilder)
-	Count  int
-}
-
-// CollectionPage is a simplified interface type
-// that can be fulfilled by either of (where required):
-// vocab.ActivityStreamsCollection
-// vocab.ActivityStreamsOrderedCollection
-type CollectionBuilder interface {
-	SetJSONLDId(vocab.JSONLDIdProperty)
-	SetActivityStreamsFirst(vocab.ActivityStreamsFirstProperty)
-	SetActivityStreamsTotalItems(i vocab.ActivityStreamsTotalItemsProperty)
-}
-
-// CollectionPageBuilder is a simplified interface type
-// that can be fulfilled by either of (where required):
-// vocab.ActivityStreamsCollectionPage
-// vocab.ActivityStreamsOrderedCollectionPage
-type CollectionPageBuilder interface {
-	SetJSONLDId(vocab.JSONLDIdProperty)
-	SetActivityStreamsPartOf(vocab.ActivityStreamsPartOfProperty)
-	SetActivityStreamsNext(vocab.ActivityStreamsNextProperty)
-	SetActivityStreamsPrev(vocab.ActivityStreamsPrevProperty)
-	SetActivityStreamsTotalItems(i vocab.ActivityStreamsTotalItemsProperty)
-}
-
-// ItemsPropertyBuilder is a simplified interface type
-// that can be fulfilled by either of (where required):
-// vocab.ActivityStreamsItemsProperty
-// vocab.ActivityStreamsOrderedItemsProperty
-type ItemsPropertyBuilder interface {
-	AppendIRI(*url.URL)
-
-	// NOTE: add more of the items-property-like interface
-	// functions here as you require them for building pages.
-}
-
-// NewASCollection builds and returns a new ActivityStreams Collection from given parameters.
-func NewASCollection(params CollectionParams) vocab.ActivityStreamsCollection {
-	collection := streams.NewActivityStreamsCollection()
-	buildCollection(collection, params, 40)
-	return collection
-}
-
-// NewASCollectionPage builds and returns a new ActivityStreams CollectionPage from given parameters (including item property appending function).
-func NewASCollectionPage(params CollectionPageParams) vocab.ActivityStreamsCollectionPage {
-	collectionPage := streams.NewActivityStreamsCollectionPage()
-	itemsProp := streams.NewActivityStreamsItemsProperty()
-	buildCollectionPage(collectionPage, itemsProp, collectionPage.SetActivityStreamsItems, params)
-	return collectionPage
-}
-
-// NewASOrderedCollection builds and returns a new ActivityStreams OrderedCollection from given parameters.
-func NewASOrderedCollection(params CollectionParams) vocab.ActivityStreamsOrderedCollection {
-	collection := streams.NewActivityStreamsOrderedCollection()
-	buildCollection(collection, params, 40)
-	return collection
-}
-
-// NewASOrderedCollectionPage builds and returns a new ActivityStreams OrderedCollectionPage from given parameters (including item property appending function).
-func NewASOrderedCollectionPage(params CollectionPageParams) vocab.ActivityStreamsOrderedCollectionPage {
-	collectionPage := streams.NewActivityStreamsOrderedCollectionPage()
-	itemsProp := streams.NewActivityStreamsOrderedItemsProperty()
-	buildCollectionPage(collectionPage, itemsProp, collectionPage.SetActivityStreamsOrderedItems, params)
-	return collectionPage
-}
-
-func buildCollection[C CollectionBuilder](collection C, params CollectionParams, pageLimit int) {
-	// Add the collection ID property.
-	idProp := streams.NewJSONLDIdProperty()
-	idProp.SetIRI(params.ID)
-	collection.SetJSONLDId(idProp)
-
-	// Add the collection totalItems count property.
-	totalItems := streams.NewActivityStreamsTotalItemsProperty()
-	totalItems.Set(params.Total)
-	collection.SetActivityStreamsTotalItems(totalItems)
-
-	// Clone the collection ID page
-	// to add first page query data.
-	firstIRI := new(url.URL)
-	*firstIRI = *params.ID
-
-	// Note that simply adding a limit signals to our
-	// endpoint to use paging (which will start at beginning).
-	limit := "limit=" + strconv.Itoa(pageLimit)
-	firstIRI.RawQuery = appendQuery(firstIRI.RawQuery, limit)
-
-	// Add the collection first IRI property.
-	first := streams.NewActivityStreamsFirstProperty()
-	first.SetIRI(firstIRI)
-	collection.SetActivityStreamsFirst(first)
-}
-
-func buildCollectionPage[C CollectionPageBuilder, I ItemsPropertyBuilder](collectionPage C, itemsProp I, setItems func(I), params CollectionPageParams) {
-	// Add the partOf property for its containing collection ID.
-	partOfProp := streams.NewActivityStreamsPartOfProperty()
-	partOfProp.SetIRI(params.ID)
-	collectionPage.SetActivityStreamsPartOf(partOfProp)
-
-	// Build the current page link IRI.
-	currentIRI := params.Current.ToLinkURL(
-		params.ID.Scheme,
-		params.ID.Host,
-		params.ID.Path,
-		params.Query,
-	)
-
-	// Add the collection ID property for
-	// the *current* collection page params.
-	idProp := streams.NewJSONLDIdProperty()
-	idProp.SetIRI(currentIRI)
-	collectionPage.SetJSONLDId(idProp)
-
-	// Build the next page link IRI.
-	nextIRI := params.Next.ToLinkURL(
-		params.ID.Scheme,
-		params.ID.Host,
-		params.ID.Path,
-		params.Query,
-	)
-
-	if nextIRI != nil {
-		// Add the collection next property for the next page.
-		nextProp := streams.NewActivityStreamsNextProperty()
-		nextProp.SetIRI(nextIRI)
-		collectionPage.SetActivityStreamsNext(nextProp)
-	}
-
-	// Build the prev page link IRI.
-	prevIRI := params.Prev.ToLinkURL(
-		params.ID.Scheme,
-		params.ID.Host,
-		params.ID.Path,
-		params.Query,
-	)
-
-	if prevIRI != nil {
-		// Add the collection prev property for the prev page.
-		prevProp := streams.NewActivityStreamsPrevProperty()
-		prevProp.SetIRI(prevIRI)
-		collectionPage.SetActivityStreamsPrev(prevProp)
-	}
-
-	// Add the collection totalItems count property.
-	totalItems := streams.NewActivityStreamsTotalItemsProperty()
-	totalItems.Set(params.Total)
-	collectionPage.SetActivityStreamsTotalItems(totalItems)
-
-	if params.Append == nil {
-		// nil check outside the for loop.
-		panic("nil params.Append function")
-	}
-
-	// Append each of the items to the provided
-	// pre-allocated items property builder type.
-	for i := 0; i < params.Count; i++ {
-		params.Append(i, itemsProp)
-	}
-
-	// Set the collection
-	// page items property.
-	setItems(itemsProp)
-}
-
-// appendQuery appends part to an existing raw
-// query with ampersand, else just returning part.
-func appendQuery(raw, part string) string {
-	if raw != "" {
-		return raw + "&" + part
-	}
-	return part
-}
